feat: add operations into create Transaction and fix date property

+++ src/main/kotlin/com/wa2/demo/dto/WalletDTO.kt
@@ -1,12 +1,13 @@
 package com.wa2.demo.dto
 
 import com.wa2.demo.entities.Wallet
+import java.math.BigDecimal
 
 
 data class WalletDTO(
     var walletId: Long? = null,
     var customer: CustomerDTO? = null,
-    var currentAmount: Long? = null,
+    var currentAmount: BigDecimal? = null,
     var payees: MutableSet<TransactionDTO> = mutableSetOf(),
     var payers: MutableSet<TransactionDTO> = mutableSetOf()
 ) {
